fix(ci): make the pipeline green
Four things the first CI run found, all of them real.

The pytest markers list had lost the slow marker, so every test job failed at
collection under --strict-markers. Strict markers are worth keeping: a typo
in a marker name should fail loudly rather than silently deselecting nothing.

The lint job installed the newest ruff while development used an older one,
so a rule that did not exist locally failed the build. The linter now comes
from the dev extra like every other tool, which puts the version in one place
and means a clean local run stays clean here.

That newer ruff was right about both findings: math.floor already returns an
int, so wrapping it in int() said nothing.

pip-audit was auditing the project itself, which has no PyPI entry to look
up, and failing on not finding it. --skip-editable limits the audit to the
dependencies, which is what the job is for.

+++ netpulse/eval/metrics.py
@@ -285,7 +285,7 @@ def percentile(values: list[float], fraction: float) -> float:
         return 0.0
     ordered = sorted(values)
     position = fraction * (len(ordered) - 1)
-    low = int(math.floor(position))
+    low = math.floor(position)
     high = min(low + 1, len(ordered) - 1)
     weight = position - low
     return ordered[low] * (1 - weight) + ordered[high] * weight
